snap

+++ source/scripts/cmake/win_pkg_cleanup.cmake
@@ -6,8 +6,13 @@ set(SUPPORT_DIR ${CMAKE_SOURCE_DIR}/support)
 set(lib_dir ${SUPPORT_DIR}/libs)
 set(include_dir ${SUPPORT_DIR}/include)
 
-cmake_path(NORMAL_PATH lib_dir)
-cmake_path(NORMAL_PATH include_dir)
+cmake_path(NORMAL_PATH lib_dir OUTPUT_VARIABLE lib_dir)
+cmake_path(NORMAL_PATH include_dir OUTPUT_VARIABLE include_dir)
 
-file(REMOVE_RECURSE ${lib_dir})
-file(REMOVE_RECURSE ${include_dir})
+if(EXISTS ${lib_dir})
+	file(REMOVE_RECURSE ${lib_dir})
+endif()
+
+if(EXISTS ${include_dir})
+	file(REMOVE_RECURSE ${include_dir})
+endif()
